pyproject.toml
[project]
name = "docs_translator"
version = "0.1.0"
description = "docs-translator using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "pygithub>=2.6.1",
]

[project.scripts]
kickoff = "docs_translator.main:kickoff"
run_crew = "docs_translator.main:kickoff"
plot = "docs_translator.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
